requirements.txt
requests==2.31.0
selenium==4.34.2
schedule==1.2.1
Pillow==10.1.0
onnxruntime==1.18.1
numpy==1.26.2
webdrivermanager_cn==2.4.0
webdriver-manager==4.0.2
mysql-connector-python==9.4.0
# python-dotenv
# python-dateutil

# 反爬虫检测绕过
undetected-chromedriver==3.5.5
selenium-stealth==1.0.6

# 点击验证码识别
ddddocr==1.4.11